rejig types for player
