+++ src/main/java/tictactoe/core/PlayerSymbol.java
@@ -1,0 +1,17 @@
+package tictactoe.core;
+
+public enum PlayerSymbol {
+    X,
+    O;
+
+    private PlayerSymbol alternate;
+
+    static {
+        X.alternate = O;
+        O.alternate = X;
+    }
+
+    public PlayerSymbol getAlternate() {
+        return alternate;
+    }
+}
